feature/P-001 implemented url hand id links to make it clearer
