+++ src/contexts/HandContext.tsx
@@ -2,6 +2,7 @@
 import { Hand } from '@/modules/hand/domain/hand';
 import { LocalStorageRepository } from '@/modules/shared/infrastructure/local_storage.repository';
 import { createContext, useState, useContext, useEffect } from 'react';
+import { useRouter } from "next/navigation";
 interface HandContextState {
     currentHandIdx: number;
     setCurrentHandIdx: (idx: number) => void;
@@ -10,6 +11,7 @@ interface HandContextState {
     nextHand: () => void;
     prevHand: () => void;
     currentHand: Hand | null;
+    loadHand: (id: string) => void;
 }
 
 export const HandContext = createContext<HandContextState>({
@@ -20,22 +22,34 @@ export const HandContext = createContext<HandContextState>({
     nextHand: () => {},
     prevHand: () => {},
     currentHand: null,
+    loadHand: () => {},
 });
 
 export function HandContextProvider({ children }: { children: React.ReactNode }) {
     const [hands, setHands] = useState<Array<Hand>>([]);
     const [currentHandIdx, setCurrentHandIdx] = useState<number>(0);
     const localStorageRepository = new LocalStorageRepository();
+    const router = useRouter();
   
     const nextHand = () => {
       if (currentHandIdx < hands.length - 1) {
+        const nextIdx = currentHandIdx + 1;
         setCurrentHandIdx(currentHandIdx + 1);
+        const next = hands[nextIdx];
+        if (next) {
+          router.push(`/hands/${next?.id}`, {scroll: true})
+        }
       }
     };
   
     const prevHand = () => {
       if (currentHandIdx > 0) {
-        setCurrentHandIdx(currentHandIdx - 1);
+        const prevIdx = currentHandIdx - 1;
+        setCurrentHandIdx(prevIdx);
+        const prev = hands[prevIdx];
+        if (prev) {
+          router.push(`/hands/${prev?.id}`, {scroll: true})
+        }
       }
     };
 
@@ -44,13 +58,17 @@ export function HandContextProvider({ children }: { children: React.ReactNode })
       if (idx > -1) {
         setCurrentHandIdx(idx)
       }
-      return hands[idx]
     }
 
     useEffect(() => {
       const existingHistory = localStorageRepository.loadHandHistory('handHistory', true);
       if (existingHistory) {
         setHands(existingHistory)
+        const urlHandId = window.location.pathname.split('/').pop(); // Get hand ID from URL
+        if (urlHandId) {
+            const idx = existingHistory.findIndex((hand) => hand.id === urlHandId);
+            if (idx > -1) setCurrentHandIdx(idx);
+        }
       } 
     }, [])
   
@@ -63,6 +81,7 @@ export function HandContextProvider({ children }: { children: React.ReactNode })
           prevHand, 
           setCurrentHandIdx,
           currentHand: hands[currentHandIdx],
+          loadHand,
       }}>
         {children}
       </HandContext.Provider>
